requirements.txt
disnake
pyyaml
dice
